Add RustyXML bench
Benches are available through features:
- xml-rs:   bench-xml-rs
- RustyXML: bench-rusty-xml
