+++ Cargo.toml
@@ -11,19 +11,15 @@ readme = "README.md"
 keywords = ["xml", "reader", "parser", "write"]
 license = "MIT"
 
+[features]
+bench-xml-rs = [ "xml-rs" ]
+bench-rusty-xml = [ "RustyXML" ]
+
 [dependencies]
 log = "0.3"
-
-[dev-dependencies]
-xml-rs = "0.3.0"
+xml-rs = { version = "0.3.0", optional = true }
+RustyXML = { version = "0.1.1", optional = true }
 
 [lib]
 bench = false
 
-[profile.bench]
-opt-level = 3
-debug = true
-rpath = false
-lto = false
-debug-assertions = false
-codegen-units = 1
